+++ SERVICE/Service.Queries/ConveniosQueryService.cs
@@ -1,6 +1,7 @@
 ﻿using Common.Collection;
 using DATA.DTOS;
 using DATA.DTOS.Updates;
+using DATA.Extensions;
 using Microsoft.EntityFrameworkCore;
 using PERSISTENCE;
 using Services.Common.Mapping;
@@ -37,12 +38,15 @@ namespace Service.Queries
                 .Where(x => convenios == null || convenios.Contains(x.IdConvenio))
                 .OrderByDescending(x => x.IdConvenio)
                 .GetPagedAsync(page, take);
-
+                if (!collection.HasItems)
+                {
+                    throw new EmptyCollectionException("No se encontró ningun Item en la Base de Datos");
+                }
                 return collection.MapTo<DataCollection<ConveniosDTO>>();
             }
             catch (Exception ex)
             {
-                throw new Exception("Error al obtener las Convenios");
+                throw new Exception("Error al obtener los Convenios");
             }
 
         }
@@ -51,25 +55,25 @@ namespace Service.Queries
         {
             try
             {
-                return (await _context.Convenios.SingleAsync(x => x.IdConvenio == id)).MapTo<ConveniosDTO>();
+                if (await _context.Convenios.FindAsync(id) == null)
+                {
+                    throw new EmptyCollectionException("Error al obtener el Convenio, el Convenio con id" + " " + id + " " + "no existe");
+                }
+                return (await _context.Convenios.FindAsync(id)).MapTo<ConveniosDTO>();
             }
             catch (Exception ex)
             {
-                if (_context.Convenios.SingleAsync(x => x.IdConvenio == id) == null)
-                {
-                    throw new Exception("Error al obtener el convenio, lel convenio con id" + " " + id + " " + "no existe");
-                }
-                throw new Exception("Error al obtener el convenio");
+                throw new Exception("Error al obtener el Convenio");
             }
 
         }
         public async Task<UpdateConvenioDTO> PutAsync(UpdateConvenioDTO Convenio, int id)
         {
-            if (_context.Convenios.SingleAsync(x => x.IdConvenio == id).Result == null)
+            if (await _context.Convenios.FindAsync(id) == null)
             {
-                throw new Exception("El convenio con id" + " " + id + " " + ",no existe");
+                throw new EmptyCollectionException("Error al actualizar el Convenio, el Convenio con id" + " " + id + " " + "no existe");
             }
-            var convenio = await _context.Convenios.SingleAsync(x => x.IdConvenio == id);
+            var convenio = await _context.Convenios.FindAsync(id);
             convenio.Descripcion = Convenio.Descripcion;
             convenio.Obs = Convenio.Obs;
 
@@ -82,17 +86,17 @@ namespace Service.Queries
             try
             {
                 var convenio = await _context.Convenios.SingleAsync(x => x.IdConvenio == id);
+                if (convenio == null)
+                {
+                    throw new EmptyCollectionException("Error al eliminar el Convenio, el Convenio con id" + " " + id + " " + "no existe");
+                }
                 _context.Convenios.Remove(convenio);
                 await _context.SaveChangesAsync();
                 return convenio.MapTo<ConveniosDTO>();
             }
             catch (Exception ex)
             {
-                if (_context.Convenios.SingleAsync(x => x.IdConvenio == id) == null)
-                {
-                    throw new Exception("Error al eliminar el Convenio, el convenio con id" + " " + id + " " + "no existe");
-                }
-                throw new Exception("Error al eliminar el convenio");
+                throw new Exception("Error al eliminar el Convenio");
             }
 
         }
