+++ SERVICE/Service.Queries/EspecialidadesQueryService.cs
@@ -1,6 +1,7 @@
 ﻿using Common.Collection;
 using DATA.DTOS;
 using DATA.DTOS.Updates;
+using DATA.Extensions;
 using Microsoft.EntityFrameworkCore;
 using PERSISTENCE;
 using Services.Common.Mapping;
@@ -37,12 +38,15 @@ namespace Service.Queries
                 .Where(x => especialidades == null || especialidades.Contains(x.IdEspecialidad))
                 .OrderByDescending(x => x.IdEspecialidad)
                 .GetPagedAsync(page, take);
-
+                if (!collection.HasItems)
+                {
+                    throw new EmptyCollectionException("No se encontró ningun Item en la Base de Datos");
+                }
                 return collection.MapTo<DataCollection<EspecialidadesDTO>>();
             }
             catch (Exception ex)
             {
-                throw new Exception("Error al obtener las especialidades");
+                throw new Exception("Error al obtener las Especialidades");
             }
 
         }
@@ -51,23 +55,23 @@ namespace Service.Queries
         {
             try
             {
-                return (await _context.Especialidades.SingleAsync(x => x.IdEspecialidad == id)).MapTo<EspecialidadesDTO>();
+                if (await _context.Especialidades.FindAsync(id) == null)
+                {
+                    throw new EmptyCollectionException("Error al obtener la Especialidad, la Especialidad con id" + " " + id + " " + "no existe");
+                }
+                return (await _context.Especialidades.FindAsync(id)).MapTo<EspecialidadesDTO>();
             }
             catch (Exception ex)
             {
-                if (_context.Especialidades.SingleAsync(x => x.IdEspecialidad == id) == null)
-                {
-                    throw new Exception("Error al obtener el especialidad, la especialidad con id" + " " + id + " " + "no existe");
-                }
                 throw new Exception("Error al obtener el especialidad");
             }
 
         }
         public async Task<UpdateEspecialidadesDTO> PutAsync(UpdateEspecialidadesDTO Especialidad, int id)
         {
-            if (_context.Especialidades.SingleAsync(x => x.IdEspecialidad == id).Result == null)
+            if (await _context.Especialidades.FindAsync(id) == null)
             {
-                throw new Exception("El especialidad con id" + " " + id + " " + ",no existe");
+                throw new EmptyCollectionException("Error al actualizar la Especialidad, la Especialidad con id" + " " + id + " " + "no existe");
             }
             var especialidad = await _context.Especialidades.SingleAsync(x => x.IdEspecialidad == id);
             especialidad.Descripcion = Especialidad.Descripcion;
@@ -82,17 +86,17 @@ namespace Service.Queries
             try
             {
                 var especialidad = await _context.Especialidades.SingleAsync(x => x.IdEspecialidad == id);
+                if (especialidad == null)
+                {
+                    throw new EmptyCollectionException("Error al eliminar la Especialidad, la Especialidad con id" + " " + id + " " + "no existe");
+                }
                 _context.Especialidades.Remove(especialidad);
                 await _context.SaveChangesAsync();
                 return especialidad.MapTo<EspecialidadesDTO>();
             }
             catch (Exception ex)
             {
-                if (_context.Especialidades.SingleAsync(x => x.IdEspecialidad == id) == null)
-                {
-                    throw new Exception("Error al eliminar especialidad, el especialidad con id" + " " + id + " " + "no existe");
-                }
-                throw new Exception("Error al eliminar el especialidad");
+                throw new Exception("Error al eliminar la Especialidad");
             }
 
         }
