+++ SERVICE/Service.Queries/ModelosQueryService.cs
@@ -10,6 +10,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using DATA.Extensions;
 
 namespace Service.Queries
 {
@@ -38,7 +39,10 @@ namespace Service.Queries
                 .Where(x => modelos == null || modelos.Contains(x.IdModelo))
                 .OrderByDescending(x => x.IdModelo)
                 .GetPagedAsync(page, take);
-
+                if (!collection.HasItems)
+                {
+                    throw new EmptyCollectionException("No se encontró ningun Item en la Base de Datos");
+                }
                 return collection.MapTo<DataCollection<ModelosDTO>>();
             }
             catch (Exception ex)
@@ -52,23 +56,23 @@ namespace Service.Queries
         {
             try
             {
-                return (await _context.Modelos.SingleAsync(x => x.IdModelo == id)).MapTo<ModelosDTO>();
+                if (await _context.Modelos.FindAsync(id) == null)
+                {
+                    throw new EmptyCollectionException("Error al obtener el Modelo, el Modelo con id" + " " + id + " " + "no existe");
+                }
+                return (await _context.Modelos.FindAsync(id)).MapTo<ModelosDTO>();
             }
             catch (Exception ex)
             {
-                if (_context.Modelos.SingleAsync(x => x.IdModelo == id) == null)
-                {
-                    throw new Exception("Error al obtener modelos, el modelo  con id" + " " + id + " " + "no existe");
-                }
-                throw new Exception("Error al obtener el modelo");
+                throw new Exception("Error al obtener el Modelo");
             }
 
         }
         public async Task<UpdateModeloDTO> PutAsync(UpdateModeloDTO Modelo, long id)
         {
-            if (_context.Modelos.SingleAsync(x => x.IdModelo == id).Result == null)
+            if (await _context.Modelos.FindAsync(id) == null)
             {
-                throw new Exception("El modelo con id" + " " + id + " " + ",no existe");
+                throw new EmptyCollectionException("Error al eliminar el Modelo, el Modelo con id" + " " + id + " " + "no existe");
             }
             var modelo = await _context.Modelos.SingleAsync(x => x.IdModelo == id);
             modelo.Modelo = Modelo.Modelo;
@@ -83,17 +87,17 @@ namespace Service.Queries
             try
             {
                 var modelo = await _context.Modelos.SingleAsync(x => x.IdModelo == id);
+                if (modelo == null)
+                {
+                    throw new EmptyCollectionException("Error al eliminar el Modelo, el Modelo con id" + " " + id + " " + "no existe");
+                }
                 _context.Modelos.Remove(modelo);
                 await _context.SaveChangesAsync();
                 return modelo.MapTo<ModelosDTO>();
             }
             catch (Exception ex)
             {
-                if (_context.Modelos.SingleAsync(x => x.IdModelo == id) == null)
-                {
-                    throw new Exception("Error al eliminar el modelo, el modelo con id" + " " + id + " " + "no existe");
-                }
-                throw new Exception("Error al eliminar el modelo");
+                throw new Exception("Error al eliminar el Modelo");
             }
 
         }
